Update patch set 7
Patch Set 7: Code-Review+1

I think for more clarification, we can add links to the groovy-provider plugin [1] which provides the runtime to be able to run this groovy script. Feel free to add this on the repository level for gerrit/plugins/scripts to link both groovy and scala runtime providers plugins.


[1] https://gerrit.googlesource.com/plugins/scripting/groovy-provider/

Patch-set: 7
Label: Code-Review=+1