Schedule groups-backends in a queue
When loading groups from an auth backend (for example LDAP), Gerrit is
subject to its latency. This might be a problem because a slow backend,
or a very large number of groups, or both, might cause the warm-cache
command to last longer than the allowed timeouts imposed by Gerrit (or a
load balancer that fronts it).

Schedule the loading of the cache asynchronously over a new dedicated
queue named: "Groups-Backend-Cache-Warmer".

This allows to terminate the SSH command in a timely manner, while
loading the cache in the background.

The status of the cache loading can be observed via the `show-queue`
command as well as the `queue/groups_backend_cache_warmer/*` metrics.

Currently the size of the queue is not configurable and it is set to 8
threads.

Change-Id: Iccbb1f05bcf04c540a68e1d70cd6be9078a50815