Introduce globalrefdb update-ref command
Allow the Gerrit admin to update refs on the global-refdb
in the same way he can update refs on the local one, using
a command similar to 'git update-ref'.

This allows to resolve split-brains situations with a
simple command line, without having to connect directly
to the global-refdb implementation console backend.

Change-Id: Ice02d4faa08ac3c275eb241d3a8249e46bede516