Add metric for sum of SHA1s
It is not possible to have a metric with a non-numeric value, so we
calculate the numeric representation of the SHA1. However, SHA1s are
too big to fit with a numeric data type, so a truncated version of the
numeric value is calculated.

The numerical calculation was taken from [1].

[1] http://www.java2s.com/example/java-utility-method/sha1/sha1hashint-string-text-d6c0e.html

Change-Id: I09bb9bda8f6e5dc655aa5b9db9bee120cc47ce51