Adapt warm-cache script for gerrit 3.7
The script is pretty old, looks like it was developed sometime in 2016
based on Gerrit 2.14 (?). As such when running it in Gerrit 3.7 some
classes are missing and therefore the commands will fail.
Change the script in a way that compiles with gerrit 3.7 version
and all exposed commands run successfully. Replace ReviewDB operations
with its successor where possible.

Change-Id: I594a32d7e4eff6dabf1c761b6dcb9342508812f9