Contribute packed-refs.lock staleness checker
Add a general utility that verify all active repositories
for any $GIT_DIR/packed-refs.lock stale files.

The presence of $GIT_DIR/packed-refs.lock prevents multiple
threads or processes from updating the packed-refs file
concurrently. However, if one of the threads or processes
die in the middle of updating the refs, the repository
packed-refs are locked forever.

The simple utility stale-packed-refs is able to understand
if a repository is active, by listening to its refs-updated events,
and reacting on any updates.

The frequency of the checks and the max age for a lock file
for being considered stale is configurable as plugin setting
in gerrit.config.

Example, for checking every 20 mins and removing all packed-refs.lock
that are older than 5 minutes:

[plugin "stale-packed-ref"]
  checkIntervalSec = 1200
  staleMaxAgeSec = 300

Change-Id: I6723cd214ec288d89c14acd2889dcb8fb7e50982