Contribute packed-refs.lock staleness checker
Add a general utility that verifies all repositories matching
a set of prefixes doesn't have a $GIT_DIR/packed-refs.lock
stale lock file.

The presence of $GIT_DIR/packed-refs.lock prevents multiple
threads or processes from updating the packed-refs file
concurrently. However, if one of the threads or processes
dies in the middle of updating the refs, the repository
packed-refs are locked forever.

The list of projects prefixes, the frequency of the checks
and the max age for a lock file for being considered stale is
configurable as plugin setting in gerrit.config.

Example, for checking every 20 mins and removing all packed-refs.lock
that are older than 5 minutes:

[plugin "stale-packed-refs"]
  checkIntervalSec = 1200
  staleMaxAgeSec = 300
  projectPrefix = project1
  projectPrefix = production/

Change-Id: I6723cd214ec288d89c14acd2889dcb8fb7e50982