Contribute packed-refs.lock staleness checker
Add a general utility that exposes the age of the packed-refs.lock
file as a metric

The presence of $GIT_DIR/packed-refs.lock prevents multiple
threads or processes from updating the packed-refs file
concurrently. However, if one of the threads or processes
dies in the middle of updating the refs, the repository
packed-refs are locked forever.

The list of projects prefixes, the frequency of the checks
and the max age for a lock file for being considered stale is
configurable as plugin setting in gerrit.config.

Change-Id: Ie8714f28e1191b1ef255d1aeffa1ca6a73a6a959