Implement ReplicationConfigOverrides.update()
The newly added `update(Config)` method to `ReplicationConfigOverrides`
allows us to update the replication configuration programmatically.

Added implementation stores remotes definitions in the "fanout" fashion
in the `refs/meta/replication` branch of `All-Projects` repository.
Other configuration options will be stored in the `replication.config`
file in the root of the branch.

If no configuration or branch is present they will be created by the
script.

Change-Id: Ifa744342536fc029a220833f8f59c9a15954ac65