Track user activity in a persistent cache
Add a new Groovy script `track-active-users-1.0.groovy` to track user
activities by entering their `id` and timestamp into
`track-active-users_cache` persistent cache.

The tracking is implemented as a side effect on a new
`GroupBackend` implementation that handles all group `UUID`'s. This way
our backend `membershipsOf(CurrentUser)` method will be called whenever
any group membership resolution is done. This gives us a single
place to get the user activities for all supported entry points (UI,
REST, SSH, git-over-http). The impact of on the user group resolution
should be neglectable and should be equal to the performance of the
persistent cache `get()` and `put()` calls. The user activity timestamp
is stored with one-minute precision.

Bear in mind, that in the case of git-over-http access to public
repositories, even when the user provides authentication details, Gerrit
will not consult any of the `GroupBackend`'s. This means that clones and
fetches of publicly visible repositories will not be counted as user
activity.

Bug: Issue Issue 327730871
Change-Id: I5827288d40bba4c6b1e7b20d9aae822db8eabc29