TrackActive: Automatically disable inactive accounts
Build on top of the `track-active-users` (and rename it to
`track-and-disable-inactive-users`) script and add the possibility
of automatically disabling inactive accounts after a preconfigured
period.

The new functionality will mark as inactive accounts that are not
present in `track-active.users_cache`. The inactivity period can be
controlled by the `users_cache` `maxAge` configuration option.

A single configuration option is supported:
* `plugin.track-active.autoDisableInterval` - set the interval for the
  user inactivity check

To prevent premature account disablement, the script will wait with the
till its first installation time (stored in
`$gerrit_site/data/track-active/install_timestamp.txt`) is older than
`maxAge` set for its cache.

To test this feature, copy the
`admin/track-and-disable-active-users.groovy` into
`$gerrit_site/plugins` and add the following configuration to
`$gerrit_site/etc/gerrit.config`:

[plugin "track-active"]
  autoDisableInterval = 1min

[cache "track-active.users_cache"]
  maxAge = 1min

Authenticate with any user (apart from default Gerrit admin), and wait
up to 4 minutes. During that time a log message about a single account
being disabled should be appended to logs and the user won't be able to
sign in anymore.

Note: When `maxAge` and `autoDisableInterval` are equal it will take up
to `2 * maxAge` time to disable the account eg. with the above TTL of
1min, when Gerrit is started at T0s, then the user signs in at T15s, the
cache entry will be invalidated at T75s; the first run of the auto-disable
will be at T60s (the account will still be active), the second run at
T120s will disable the account. That's the reason why during testing one
need to wait up to 4 minutes for the account to be disabled.

Bug: Issue: 327730872
Change-Id: I34974d842d6f73784d43932b587a6461eaf2ffae