Do not automatically disable non-interactive users
The "Service Accounts" (aka. non-interactive users) should not be
automatically disabled by the script. Those accounts are used for CI
and other integrations and may be dormant for some time, we should not
increase Gerrit Admin workload by automatically disabling them. This
could also lead to frustration when a "paused" service is back online
but it doesn't work because their user was blocked.

Additionally, bumps the script version number to 1.1, which results in
file rename.

Bug: Issue 338071090
Change-Id: Ia09d000c72be062b8a00fa634e3ef75c85aab3bc