Fix `unexpected token: &&` and missing parenthesis
Upon loading of the plugin Gerrit was throwing an excepion because the
&& operator is not handled correctly in some cases when it starts
a new line. [1]

The logic to find all active accounts also has been fixed by adding
a parenthesis that was missing when checking if the user `isServiceUser`
before adding it to the internal cache.

This prevented the loading of the plugin, introduced in change 424058.

[1]: https://issues.apache.org/jira/browse/GROOVY-7218

Bug: Issue 370792796
Change-Id: I88f2cf40aa6658c0a91e35f21bbadd49ffcddcfe