Add `active_users` metric
This change exposes a new metric called `active_users`.
This metric indicates the number of active users; a user is considered
active when its inactivity period is not greater than the value
defined in the `cache."@PLUGIN@.users_cache".maxAge`.

The metric uses the `users_cache` of the
track-and-disable-inactive-users-1.3.groovy plugin to allow exposure
of this value to the various monitoring and observability stacks.

Bug: Issue 370937461
Change-Id: Iaa1a78a7435f80508eeeceb6df762f6c789a0bdc