Add `is_gc_running_per_project_<repo_name>` metric
Introduce a Gerrit metric to track if a GC process is running
for specific projects. The metric checks for the presence of a
`gc.pid` file in the repository and ensures the file is not older
than the configured staleness period.

By default, the staleness period is 12 hours and can be customized
using the configuration:
`plugins.repo-repack-tracker.considerGcStaleAfter`.

Change-Id: Ib19541280a913eece34b96975412673c932a7370