Consider the current logged-in user for active users tracking
Before this change, the user tracked for freshening the set
of active users was the one checked for group ownership, which
may or may not be the user actively using Gerrit.

A user admin scanning the permissions of all users could therefore
trigger implicitly the including of all accounts as active, which
would not be the case.

Fetch the current thread active user and use that account id
for tracking the activity, rather than the account passed as
paramemter for membership check.

Change-Id: Id38e0bb5fd6f41effde9c76055c2f1daeab3a983